# openapi.yaml
openapi: 3.1.0
info:
  title: DNA Futsal Scraper API
  version: 1.0.0
  description: API de coleta controlada dos dados públicos do portal de eventos da FPFS.
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
paths:
  /api/v1/division:
    get:
      summary: Lista as divisÃµes disponÃ­veis para o front-end
      parameters:
        - $ref: '#/components/parameters/OptionalSeason'
      responses:
        '200':
          description: DivisÃµes encontradas nos filtros do portal
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/divisions:
    get:
      summary: Alias de /api/v1/division
      parameters:
        - $ref: '#/components/parameters/OptionalSeason'
      responses:
        '200':
          description: DivisÃµes encontradas nos filtros do portal
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        default: { $ref: '#/components/responses/Problem' }
  /api/v1/categories:
    get:
      summary: Lista as categorias disponÃ­veis para o front-end
      parameters:
        - $ref: '#/components/parameters/OptionalSeason'
        - name: division
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Categorias encontradas nos filtros do portal
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
        default: { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Opcional quando DNA_FUTSAL_API_KEY estiver vazia.
  parameters:
...
    OptionalSeason:
      name: season
      in: query
      required: false
      description: Quando omitida, usa o ano atual do servidor.
      schema: { type: integer, minimum: 2016, maximum: 2100 }
  responses:
    Problem:
      description: Erro no formato RFC 9457 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
...
    ProblemDetail:
      type: object
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        timestamp: { type: string, format: date-time }